well shit

+++ Stage 5/7.cpp
@@ -16,7 +16,7 @@ int main(){
     string num = "264153";
     vector<int> ans;
     convert(num, ans);
-    for (int i = 0; i < ans.size(); ++i){
+    for (int i = 0; i < (int) ans.size(); ++i){
         cout << ans[i] << '\t';
     }
     cout << '\n';
